+++ src/fonduer/learning/__init__.py
@@ -1,8 +1,0 @@
-from fonduer.learning.disc_models.logistic_regression import LogisticRegression
-from fonduer.learning.disc_models.lstm import LSTM
-from fonduer.learning.disc_models.sparse_logistic_regression import (
-    SparseLogisticRegression,
-)
-from fonduer.learning.disc_models.sparse_lstm import SparseLSTM
-
-__all__ = ["LogisticRegression", "LSTM", "SparseLogisticRegression", "SparseLSTM"]